#218 [rename] 공통 로깅 클래스 파일의 위치를 변경한다.
